Removed rag.py, system.py, rag_visualization.png and updated main.py
